I improved the login performance by running some of the calls concurrently.

I also added detailed timing info to ease future debugging:

```
login
	decrypt: 616.410888671875ms
	swap-db: 6.993896484375ms
	watch-currency-history: 1352.623046875ms
	watch-cmc: 1872.888671875ms
	enable-currencies: 17733.56201171875ms
	watch-currencies: 13172.137939453125ms
	login-total: 32655.123046875ms
```

Notice how `enable-currencies` and `watch-currencies` are very slow. Both of which are calling Marketmaker...