@lukechilds 2018-08-14
Brilliant!

This is actually working towards a similar idea I had where the entire login process would be wrapped in an event emitter that would emit textual progress events.

So instead of just showing the static `Logging in...` text for ages we can show the latest text from the event emitter so it would cycle through messages like:

```
Decrypting Portfolio...
Building SwapDB...
Fetching Fiat Prices...
Fetching Historical Prices...
Enabling KMD...
Enabling CHIPS...
Enabling ETH...
Enabling MYTH...
Getting Portfolio Balances...
```

That would give the user a bit more insight as to what's going on rather than maybe them thinking something's crashed if the view doesn't change for a while.

It could also be useful for debugging if we notice enabling a certain currency always takes a while.

Anyway, that's just another idea I had, beyond the scope of this PR, but good to see this refactor is working towards that kind of structure.