@lukechilds 2018-08-14
And regarding poor mm performance, it processes each request serially and blocks everything else (as well as getting blocked by anything else that it's currently doing internally).

I believe it also queries Electrum servers pretty inefficiently by opening a new connection for each request as opposed to persisting the connection and getting pushed live updates on balances.

Hopefully, we can push to get this resolved in mmv2.

// cc @ArtemGr @artemii235