@sindresorhus 2018-08-14
> This is actually working towards a similar idea I had where the entire login process would be wrapped in an event emitter that would emit textual progress events.

I don't think this is something we should expose to the user. I would rather use the time teaching them how HyperDEX works, by showing an intro the first time, and later some random tips.